+++ ChangeLog.md
@@ -11,6 +11,7 @@
     - Obsolete Modal.buttonEnable() and Modal.setButtons()
     - Review and improve button management, notably changing the 'md-click' event data (todo #28, todo #36)
     - Use 'this' instead of Temmplate.currentData() inside of helpers and events
+    - Restore Modal.target() as a pure getter
 
 ### 1.7.3